pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "simboba"
version = "0.2.0"
description = "A lightweight tool for generating annotated eval datasets and running LLM-as-judge evaluations"
readme = "README.md"
requires-python = ">=3.9"
license = "MIT"
authors = [
    { name = "Krishna Nandakumar", email = "krishna@getaide.ai" }
]
keywords = ["llm", "eval", "evaluation", "testing", "ai", "judge"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Environment :: Web Environment",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Testing",
]
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "pydantic>=2.0.0",
    "click>=8.0.0",
    "litellm>=1.0.0",
    "python-multipart>=0.0.6",
    "pypdf>=3.0.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "httpx>=0.25.0",
]

[project.urls]
Homepage = "https://github.com/ntkris/simboba"
Repository = "https://github.com/ntkris/simboba"
Issues = "https://github.com/ntkris/simboba/issues"

[project.scripts]
boba = "simboba.cli:main"
simboba = "simboba.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["simboba"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v"
